backend/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
pillow
sqlalchemy
requests
